+++ test/App/Fossa/Config/AnalyzeSpec.hs
@@ -81,7 +81,7 @@ spec = do
       it' "should set correct filters when --exclude-manifest-strategies is set" $ do
         let cfgFile = Nothing
         cliOpts <- parseArgString cliParser "--exclude-manifest-strategies"
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           ([], excludedTargets) -> length excludedTargets `shouldBe'` numberOfStrategies
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -89,7 +89,7 @@ spec = do
       it' "should set correct filters when only filter is set" $ do
         let cfgFile = Nothing
         cliOpts <- parseArgString cliParser "--only-target npm"
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           (includedTargets, []) -> includedTargets `shouldBe'` [TypeTarget "npm"]
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -97,7 +97,7 @@ spec = do
       it' "should have --exclude-manifest-strategies override only/exclude filters" $ do
         let cfgFile = Nothing
         cliOpts <- parseArgString cliParser "--exclude-manifest-strategies --only-target npm"
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           ([], excludedTargets) -> length excludedTargets `shouldBe'` numberOfStrategies
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -106,7 +106,7 @@ spec = do
       it' "should set correct filters when targets.excludeManifestStrategies is set" $ do
         let cfgFile = Just $ configFileWithTargets [] [] True
         cliOpts <- parseArgString cliParser ""
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           ([], excludedTargets) -> length excludedTargets `shouldBe'` numberOfStrategies
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -114,7 +114,7 @@ spec = do
       it' "should set correct filters when targets.only is set" $ do
         let cfgFile = Just $ configFileWithTargets ["npm"] [] False
         cliOpts <- parseArgString cliParser ""
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           (includedTargets, []) -> includedTargets `shouldBe'` [TypeTarget "npm"]
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -122,7 +122,7 @@ spec = do
       it' "should have targets.excludeManifestStrategies override only/exclude filters" $ do
         let cfgFile = Just $ configFileWithTargets ["npm"] [] True
         cliOpts <- parseArgString cliParser ""
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           ([], excludedTargets) -> length excludedTargets `shouldBe'` numberOfStrategies
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -131,7 +131,7 @@ spec = do
       it' "should ignore config file specifying targets.excludeManifestStrategies and just use CLI options" $ do
         let cfgFile = Just $ configFileWithTargets [] [] True
         cliOpts <- parseArgString cliParser "--only-target npm"
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           (includedTargets, []) -> includedTargets `shouldBe'` [TypeTarget "npm"]
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)
@@ -139,7 +139,7 @@ spec = do
       it' "should ignore config file specifying targets.only and just use CLI options" $ do
         let cfgFile = Just $ configFileWithTargets ["npm"] [] False
         cliOpts <- parseArgString cliParser "--only-target gomod"
-        filters <- filterSet <$> mergeOpts cfgFile envVars cliOpts
+        filters <- filterSet <$> mergeOpts Nothing cfgFile envVars cliOpts
         case (combinedTargets $ includeFilters filters, combinedTargets $ excludeFilters filters) of
           (includedTargets, []) -> includedTargets `shouldBe'` [TypeTarget "gomod"]
           _ -> expectationFailure' ("Incorrect filters applied. Got " ++ show filters)